Code organization and jest added
Code re-organized to modularize for easier testing
Minor tests added
added jest and set all the package.json defaults
