+++ src/demo.json
@@ -2,16 +2,6 @@
   "name": "Normal Distribution Calculator",
   "qset": {
     "version": 1,
-    "data": {
-      "items": [
-        {
-          "materiaType": "question",
-          "id": null,
-          "type": "MC",
-          "questions": [],
-          "answers": []
-        }
-      ]
-    }
+    "data": {}
   }
 }
