+++ src/player.html
@@ -29,6 +29,6 @@
 			</div>
 			<div class="info">&mu;=E(X) = <span id="meanOut">0</span>&nbsp; &sigma; = SD(X) = <span id="stddevOut">1</span>&nbsp; &sigma;&sup2; = Var(X) = <span id="varOut">1</span></div>
 		</div>
-		<script src="controllers/player.js"></script>
+		<script src="player.js"></script>
 	</body>
 </html>
