+++ webpack.config.js
@@ -1,26 +1,22 @@
 const path = require('path');
-
 const widgetWebpack = require('materia-widget-development-kit/webpack-widget')
-
 const rules = widgetWebpack.getDefaultRules()
 const entries = widgetWebpack.getDefaultEntries();
 
-// cusomize the config
+// remove some default entries
 delete entries['creator.js'];
 delete entries['creator.css'];
 delete entries['player.js'];
 
-entries['controllers/player.js'] = [path.join(__dirname, 'src', 'controllers', 'player.js')];
+// add our player controller
+entries['player.js'] = [path.join(__dirname, 'src', 'player.js')];
 
 // All JS passed through Babel
 babelJS = {
   test: /\.js$/i,
   exclude: /node_modules/,
   use: {
-    loader: 'babel-loader',
-    options: {
-      presets: ['env']
-    }
+    loader: 'babel-loader' // note: using .babelrc for presets
   }
 };
 
@@ -39,5 +35,3 @@ let options = {
 };
 
 module.exports = widgetWebpack.getLegacyWidgetBuildConfig(options);;
-
-
